backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6
pillow==9.5.0
pydantic==1.10.12